@@ -1,8 +1,6 @@
 module github.com/coinbase/cb-mpc-go
 
-go 1.23.0
-
-toolchain go1.23.12
+go 1.25.2
 
 require golang.org/x/tools v0.31.0
 
